Rename activity.js to Activity.js
