+++ pages/activity/Activity.js
