+++ pages/activity/index.js
@@ -1,3 +1,3 @@
-import Activity from './activity';
+import Activity from './Activity';
 
 export default Activity;
